Display subsystem and camera snap
The Display subsystem sets the screen to an optimal resolution based on 640X360 and the display aspect ratio, and can scale the screen.

The Camera rotation can now snap or slide to 45degrees. when the camera stops being rotated, with snap, it will move toward the nearest multiple of 45, with slide, it will continue moving towards the next 45. plus the logic is now on the camera struct